Mark cstr_to_string as unsafe
In Rust, having an unsafe interior conventionally means that you are
guaranteeing safety at the boundary. However, that's not the case for
cstr_to_string which is taking a random pointer as input. Thankfully,
this doesn't functionally change anything anyway since it was already
always called from an unsafe context.